♻️ Move search page into its own module
